[S07P12C201-264] [BE-박상현] fix : post 저장시 post id return

+++ backend/restapiserver/src/main/java/com/JJP/restapiserver/service/post/PostService.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface PostService {
 
     // 스테이지 포스트 등록
-    int savePost(PostSaveRequestDto postSaveRequestDto, Long member_id);
+    Long savePost(PostSaveRequestDto postSaveRequestDto, Long member_id);
 
     // 스테이지 포스트 수정
     int updatePost(PostUpdateRequestDto postUpdateRequestDto, Long member_id);
